2023/11/20
Today progress I did some changing in the guidelines where we could see the material that can be recycled, and the material that cannot be recycled. I add a page for the material identification (AI) , some change in the profile page ( chick on it because of the database link) , also in the info page I didn't know what to do I added some information how much it's important to recycle and the Saudi vision 2030 please could you fix the link between the pages in the request_page.dart when you did, the import still not working thank you so much buddy.